fixed a few problemos

+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Community/CommunityLiveStatus.cs
@@ -7,27 +7,27 @@ namespace Community
 {
     public class CommunityLiveStatus
     {
-        public string dateStatusUpdated;
-        public string url;
-        public string partnershipIdentifier;
-        public Partnerships.PartnershipType partnershipType;
-        public string thumbnail;
-        public string thumbnailSmall;
-        public string thumbnailLarge;
-        public int destinyCharacterId;
-        public User.UserInfoCard userInfo;
-        public int currentActivityHash;
-        public string dateLastPlayed;
-        public string dateStreamStarted;
-        public string locale;
-        public int currentViewers;
-        public int followers;
-        public int overallViewers;
-        public bool isFeatured;
-        public string title;
-        public int activityModeHash;
-        public string dateFeatured;
-        public decimal trendingValue;
-        public bool isSubscribable;
-    }
+		public string dateStatusUpdated;
+		public string url;
+		public string partnershipIdentifier;
+public Partnerships.PartnershipType partnershipType;
+		public string thumbnail;
+		public string thumbnailSmall;
+		public string thumbnailLarge;
+		public int destinyCharacterId;
+public User.UserInfoCard userInfo;
+		public ulong currentActivityHash;
+		public string dateLastPlayed;
+		public string dateStreamStarted;
+		public string locale;
+		public int currentViewers;
+		public int followers;
+		public int overallViewers;
+		public bool isFeatured;
+		public string title;
+		public ulong activityModeHash;
+		public string dateFeatured;
+		public decimal trendingValue;
+		public bool isSubscribable;
+	}
 }
