+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Character/DestinyCharacterCustomization.cs
@@ -7,17 +7,17 @@ namespace Destiny.Character
 {
     public class DestinyCharacterCustomization
     {
-        public int personality;
-        public int face;
-        public int skinColor;
-        public int lipColor;
-        public int eyeColor;
-        public UInt32[] hairColors;
-        public UInt32[] featureColors;
-        public int decalColor;
-        public bool wearHelmet;
-        public int hairIndex;
-        public int featureIndex;
-        public int decalIndex;
-    }
+		public ulong personality;
+		public ulong face;
+		public ulong skinColor;
+		public ulong lipColor;
+		public ulong eyeColor;
+		public UInt32[] hairColors;
+		public UInt32[] featureColors;
+		public ulong decalColor;
+		public bool wearHelmet;
+		public int hairIndex;
+		public int featureIndex;
+		public int decalIndex;
+	}
 }
