+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions.Director/DestinyActivityGraphDisplayProgressionDefinition.cs
@@ -7,7 +7,7 @@ namespace Destiny.Definitions.Director
 {
     public class DestinyActivityGraphDisplayProgressionDefinition
     {
-        public int id;
-        public int progressionHash;
-    }
+		public ulong id;
+		public ulong progressionHash;
+	}
 }
