+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions.Director/DestinyActivityGraphNodeDefinition.cs
@@ -7,8 +7,8 @@ namespace Destiny.Definitions.Director
 {
     public class DestinyActivityGraphNodeDefinition
     {
-        public int nodeId;
-        public Destiny.Definitions.Director.DestinyActivityGraphNodeFeaturingStateDefinition[] featuringStates;
-        public Destiny.Definitions.Director.DestinyActivityGraphNodeActivityDefinition[] activities;
-    }
+		public ulong nodeId;
+		public Destiny.Definitions.Director.DestinyActivityGraphNodeFeaturingStateDefinition[] featuringStates;
+		public Destiny.Definitions.Director.DestinyActivityGraphNodeActivityDefinition[] activities;
+	}
 }
