+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions.Director/DestinyLinkedGraphDefinition.cs
@@ -7,11 +7,11 @@ namespace Destiny.Definitions.Director
 {
     public class DestinyLinkedGraphDefinition
     {
-        public string description;
-        public string name;
-        public Destiny.Definitions.DestinyUnlockExpressionDefinition unlockExpression;
-        public int linkedGraphId;
-        public Destiny.Definitions.Director.DestinyLinkedGraphEntryDefinition[] linkedGraphs;
-        public string overview;
-    }
+		public string description;
+		public string name;
+public Destiny.Definitions.DestinyUnlockExpressionDefinition unlockExpression;
+		public ulong linkedGraphId;
+		public Destiny.Definitions.Director.DestinyLinkedGraphEntryDefinition[] linkedGraphs;
+		public string overview;
+	}
 }
