+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions.Lore/DestinyLoreDefinition.cs
@@ -7,10 +7,10 @@ namespace Destiny.Definitions.Lore
 {
     public class DestinyLoreDefinition
     {
-        public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
-        public string subtitle;
-        public int hash;
-        public int index;
-        public bool redacted;
-    }
+public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
+		public string subtitle;
+		public ulong hash;
+		public int index;
+		public bool redacted;
+	}
 }
