+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions.Milestones/DestinyMilestoneDefinition.cs
@@ -7,16 +7,16 @@ namespace Destiny.Definitions.Milestones
 {
     public class DestinyMilestoneDefinition
     {
-        public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
-        public string image;
-        public bool recruitable;
-        public string friendlyName;
-        public bool showInExplorer;
-        public bool hasPredictableDates;
-        public Destiny.Definitions.Milestones.DestinyMilestoneVendorDefinition[] vendors;
-        public bool isInGameMilestone;
-        public int hash;
-        public int index;
-        public bool redacted;
-    }
+public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
+		public string image;
+		public bool recruitable;
+		public string friendlyName;
+		public bool showInExplorer;
+		public bool hasPredictableDates;
+		public Destiny.Definitions.Milestones.DestinyMilestoneVendorDefinition[] vendors;
+		public bool isInGameMilestone;
+		public ulong hash;
+		public int index;
+		public bool redacted;
+	}
 }
