+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions.Milestones/DestinyMilestoneQuestRewardItem.cs
@@ -7,10 +7,10 @@ namespace Destiny.Definitions.Milestones
 {
     public class DestinyMilestoneQuestRewardItem
     {
-        public int vendorHash;
-        public int vendorItemIndex;
-        public int itemHash;
-        public int itemInstanceId;
-        public int quantity;
-    }
+		public ulong vendorHash;
+		public int vendorItemIndex;
+		public ulong itemHash;
+		public int itemInstanceId;
+		public int quantity;
+	}
 }
