+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions.Milestones/DestinyMilestoneRewardEntryDefinition.cs
@@ -7,10 +7,10 @@ namespace Destiny.Definitions.Milestones
 {
     public class DestinyMilestoneRewardEntryDefinition
     {
-        public int rewardEntryHash;
-        public string rewardEntryIdentifier;
-        public Destiny.DestinyItemQuantity[] items;
-        public int vendorHash;
-        public int order;
-    }
+		public ulong rewardEntryHash;
+		public string rewardEntryIdentifier;
+		public Destiny.DestinyItemQuantity[] items;
+		public ulong vendorHash;
+		public int order;
+	}
 }
