+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions.Sockets/DestinySocketTypeDefinition.cs
@@ -7,11 +7,11 @@ namespace Destiny.Definitions.Sockets
 {
     public class DestinySocketTypeDefinition
     {
-        public Destiny.Definitions.Sockets.DestinyPlugWhitelistEntryDefinition[] plugWhitelist;
-        public int socketCategoryHash;
-        public Destiny.DestinySocketVisibility visibility;
-        public int hash;
-        public int index;
-        public bool redacted;
-    }
+		public Destiny.Definitions.Sockets.DestinyPlugWhitelistEntryDefinition[] plugWhitelist;
+		public ulong socketCategoryHash;
+public Destiny.DestinySocketVisibility visibility;
+		public ulong hash;
+		public int index;
+		public bool redacted;
+	}
 }
