+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions/DestinyActivityModeDefinition.cs
@@ -7,16 +7,16 @@ namespace Destiny.Definitions
 {
     public class DestinyActivityModeDefinition
     {
-        public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
-        public string pgcrImage;
-        public Destiny.HistoricalStats.Definitions.DestinyActivityModeType modeType;
-        public Destiny.DestinyActivityModeCategory activityModeCategory;
-        public UInt32[] parentHashes;
-        public string friendlyName;
-        public bool display;
-        public int order;
-        public int hash;
-        public int index;
-        public bool redacted;
-    }
+public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
+		public string pgcrImage;
+public Destiny.HistoricalStats.Definitions.DestinyActivityModeType modeType;
+public Destiny.DestinyActivityModeCategory activityModeCategory;
+		public UInt32[] parentHashes;
+		public string friendlyName;
+		public bool display;
+		public int order;
+		public ulong hash;
+		public int index;
+		public bool redacted;
+	}
 }
