+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions/DestinyClassDefinition.cs
@@ -7,10 +7,10 @@ namespace Destiny.Definitions
 {
     public class DestinyClassDefinition
     {
-        public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
-        public int mentorVendorHash;
-        public int hash;
-        public int index;
-        public bool redacted;
-    }
+public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
+		public ulong mentorVendorHash;
+		public ulong hash;
+		public int index;
+		public bool redacted;
+	}
 }
