+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions/DestinyDestinationDefinition.cs
@@ -7,14 +7,14 @@ namespace Destiny.Definitions
 {
     public class DestinyDestinationDefinition
     {
-        public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
-        public int placeHash;
-        public int defaultFreeroamActivityHash;
-        public Destiny.Definitions.DestinyActivityGraphListEntryDefinition[] activityGraphEntries;
-        public Destiny.Definitions.DestinyDestinationBubbleSettingDefinition[] bubbleSettings;
-        public Destiny.Definitions.DestinyBubbleDefinition[] bubbles;
-        public int hash;
-        public int index;
-        public bool redacted;
-    }
+public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
+		public ulong placeHash;
+		public ulong defaultFreeroamActivityHash;
+		public Destiny.Definitions.DestinyActivityGraphListEntryDefinition[] activityGraphEntries;
+		public Destiny.Definitions.DestinyDestinationBubbleSettingDefinition[] bubbleSettings;
+		public Destiny.Definitions.DestinyBubbleDefinition[] bubbles;
+		public ulong hash;
+		public int index;
+		public bool redacted;
+	}
 }
