+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions/DestinyFactionDefinition.cs
@@ -7,10 +7,10 @@ namespace Destiny.Definitions
 {
     public class DestinyFactionDefinition
     {
-        public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
-        public int progressionHash;
-        public int hash;
-        public int index;
-        public bool redacted;
-    }
+public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
+		public ulong progressionHash;
+		public ulong hash;
+		public int index;
+		public bool redacted;
+	}
 }
