+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions/DestinyGenderDefinition.cs
@@ -7,9 +7,9 @@ namespace Destiny.Definitions
 {
     public class DestinyGenderDefinition
     {
-        public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
-        public int hash;
-        public int index;
-        public bool redacted;
-    }
+public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
+		public ulong hash;
+		public int index;
+		public bool redacted;
+	}
 }
