+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions/DestinyInventoryBucketDefinition.cs
@@ -7,14 +7,14 @@ namespace Destiny.Definitions
 {
     public class DestinyInventoryBucketDefinition
     {
-        public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
-        public int bucketOrder;
-        public int itemCount;
-        public bool hasTransferDestination;
-        public bool enabled;
-        public bool fifo;
-        public int hash;
-        public int index;
-        public bool redacted;
-    }
+public Destiny.Definitions.Common.DestinyDisplayPropertiesDefinition displayProperties;
+		public int bucketOrder;
+		public int itemCount;
+		public bool hasTransferDestination;
+		public bool enabled;
+		public bool fifo;
+		public ulong hash;
+		public int index;
+		public bool redacted;
+	}
 }
