+++ DestinySharpRewritten/DestinySharpRewritten/Entities/Destiny.Definitions/DestinyInventoryItemStatDefinition.cs
@@ -7,9 +7,9 @@ namespace Destiny.Definitions
 {
     public class DestinyInventoryItemStatDefinition
     {
-        public int statHash;
-        public int value;
-        public int minimum;
-        public int maximum;
-    }
+		public ulong statHash;
+		public int value;
+		public int minimum;
+		public int maximum;
+	}
 }
